@@ -7,6 +7,7 @@
 		"dev": "webpack-dev-server --colors --env.mode development",
 		"build": "npm run clean | webpack --env.mode production",
 		"build:analyze": "yarn build -- --env.presets analyze",
+		"lint": "eslint --fix --ext .tsx,.ts .",
 		"test": "jest",
 		"clean": "rm -rf ./dist"
 	},
@@ -22,56 +23,58 @@
 		"webpack"
 	],
 	"devDependencies": {
-		"@babel/core": "^7.8.3",
-		"@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
-		"@babel/plugin-proposal-optional-chaining": "^7.8.3",
-		"@babel/plugin-transform-runtime": "^7.8.3",
-		"@babel/preset-env": "^7.8.3",
-		"@babel/preset-react": "^7.8.3",
-		"@babel/preset-typescript": "^7.8.3",
-		"@types/styled-components": "^4.4.2",
-		"@types/webpack-env": "^1.15.0",
-		"@typescript-eslint/eslint-plugin": "^2.16.0",
-		"@typescript-eslint/parser": "^2.16.0",
-		"babel-jest": "24.9.0",
-		"babel-loader": "8.0.6",
-		"babel-plugin-styled-components": "^1.10.6",
-		"dotenv": "^8.2.0",
+		"@babel/core": "^7.9.0",
+		"@babel/plugin-transform-runtime": "^7.9.0",
+		"@babel/preset-env": "^7.9.0",
+		"@babel/preset-react": "^7.9.4",
+		"@babel/preset-typescript": "^7.9.0",
+		"@types/styled-components": "^5.0.1",
+		"@types/webpack-env": "^1.15.1",
+		"@typescript-eslint/eslint-plugin": "^2.26.0",
+		"@typescript-eslint/parser": "^2.26.0",
+		"babel-jest": "25.2.6",
+		"babel-loader": "8.1.0",
+		"babel-plugin-styled-components": "^1.10.7",
 		"eslint": "^6.8.0",
+		"eslint-config-prettier": "^6.10.1",
 		"eslint-config-react": "1.1.7",
-		"eslint-config-standard": "^14.1.0",
-		"eslint-loader": "3.0.3",
-		"eslint-plugin-import": "^2.20.0",
+		"eslint-config-standard": "^14.1.1",
+		"eslint-loader": "4.0.0",
+		"eslint-plugin-import": "^2.20.2",
 		"eslint-plugin-jsx-a11y": "6.2.3",
-		"eslint-plugin-node": "^11.0.0",
+		"eslint-plugin-node": "^11.1.0",
+		"eslint-plugin-prettier": "^3.1.2",
 		"eslint-plugin-promise": "^4.2.1",
-		"eslint-plugin-react": "7.18.0",
+		"eslint-plugin-react": "7.19.0",
 		"eslint-plugin-standard": "^4.0.1",
-		"file-loader": "5.0.2",
-		"html-webpack-plugin": "3.2.0",
+		"file-loader": "6.0.0",
+		"html-webpack-plugin": "4.0.4",
 		"image-webpack-loader": "6.0.0",
-		"jest": "24.9.0",
+		"jest": "25.2.7",
 		"json-loader": "0.5.7",
-		"react-test-renderer": "16.12.0",
-		"typescript": "^3.7.5",
-		"webpack": "4.41.5",
-		"webpack-bundle-analyzer": "3.6.0",
-		"webpack-cli": "3.3.10",
-		"webpack-dev-server": "3.10.1",
+		"prettier": "^2.0.2",
+		"react-test-renderer": "16.13.1",
+		"typescript": "^3.8.3",
+		"webpack": "4.42.1",
+		"webpack-bundle-analyzer": "3.6.1",
+		"webpack-cli": "3.3.11",
+		"webpack-dev-server": "3.10.3",
 		"webpack-merge": "4.2.2"
 	},
 	"dependencies": {
-		"@apollo/react-hooks": "^3.1.3",
-		"@hot-loader/react-dom": "^16.11.0",
-		"@types/react": "^16.9.17",
-		"@types/react-dom": "^16.9.4",
+		"@apollo/react-hooks": "^3.1.4",
+		"@hot-loader/react-dom": "^16.13.0",
+		"@types/react": "^16.9.32",
+		"@types/react-dom": "^16.9.6",
 		"@types/react-router-dom": "^5.1.3",
 		"apollo-boost": "^0.4.7",
-		"graphql": "^14.5.8",
-		"react": "16.12.0",
-		"react-dom": "16.12.0",
-		"react-hot-loader": "^4.12.18",
-		"react-router-dom": "^5.1.2",
-		"styled-components": "^5.0.0"
+		"dotenv-webpack": "^1.7.0",
+		"graphql": "^15.0.0",
+		"react": "16.13.1",
+		"react-dom": "16.13.1",
+		"react-hot-loader": "^4.12.20",
+		"react-router": "^6.0.0-alpha.2",
+		"react-router-dom": "^6.0.0-alpha.2",
+		"styled-components": "^5.0.1"
 	}
 }
