@@ -1,12 +1,9 @@
-
 const webpackMerge = require('webpack-merge')
 
-const applyPresets = (env = {presets: []}) => {
+const applyPresets = (env = { presets: [] }) => {
 	const presets = env.presets || []
 	const mergedPresets = [].concat(...[presets])
-	const mergedConfigs = mergedPresets.map(presetName =>
-		require(`./presets/webpack.${presetName}`)(env)
-	)
+	const mergedConfigs = mergedPresets.map((presetName) => require(`./presets/webpack.${presetName}`)(env))
 
 	return webpackMerge({}, ...mergedConfigs)
 }
