@@ -1,17 +1,16 @@
-
 import React, { useEffect, useState } from 'react'
 import styled, { ThemeProvider, createGlobalStyle } from 'styled-components'
 
 import { CustomPlayer } from './components/CustomPlayer'
 
 import { hot } from 'react-hot-loader/root'
-import { Routes } from './Routes'
+import { Router } from './Routes'
 import { setAccessToken } from './accessToken'
 import { PlayerContext } from './UseContext'
 import { refreshToken } from './helpers'
 import { Episode } from './models/models'
 import { useMutation, useLazyQuery } from '@apollo/react-hooks'
-import { GET_USER, SET_THEME } from './query/query'
+import { USER_SETTINGS, SET_THEME } from './query/user_query'
 import useTimeout from './hooks/useTimeout'
 
 const GlobalStyle = createGlobalStyle`	
@@ -21,6 +20,10 @@ const GlobalStyle = createGlobalStyle`
 	body {
 		background-color: #F7F7F7;
 		margin: 0;
+	}
+	ul {
+		list-style-type: none;
+		padding-left: 0;
 	}
 `
 
@@ -43,37 +46,38 @@ const themeDark = {
 }
 
 const Wrapper = styled.div`
-  margin-right: auto;
-  margin-left: auto;
-  max-width: 600px;
-  font-family: 'Avenir';
-  background-color: ${(props): string => props.theme.bg};
-  color: ${(props): string => props.theme.fontColor};
-  min-height: 100vh;
+	margin-right: auto;
+	margin-left: auto;
+	max-width: 600px;
+	font-family: 'Avenir';
+	background-color: ${(props): string => props.theme.bg};
+	color: ${(props): string => props.theme.fontColor};
+	min-height: 100vh;
 `
 
 interface PlayerState {
-	episode: Episode | null;
-	isPlayerVisible: boolean;
+	episode: Episode | null
+	isPlayerVisible: boolean
 }
 
-const App: React.FC = (): JSX.Element => {
+const App = (): JSX.Element => {
 	const [user, setUser] = useState<string>('')
 	const [{ episode, isPlayerVisible }, setPlayerValues] = useState<PlayerState>({
 		episode: null,
 		isPlayerVisible: false
 	})
 
-	const [loadUser, { data: userTheme }] = useLazyQuery(GET_USER)
+	const [loadUser, { data: userTheme }] = useLazyQuery(USER_SETTINGS)
 	const [setTheme] = useMutation(SET_THEME)
 
-	useTimeout((): any =>
-		refreshToken()
-			.then(data => {
+	useTimeout(
+		() =>
+			refreshToken().then((data) => {
 				setAccessToken(data.accessToken)
 				setUser(data.accessToken)
-			})
-	, 840000)
+			}),
+		780000
+	)
 
 	const handleUser = (value: string): void => {
 		setUser(value)
@@ -84,16 +88,15 @@ const App: React.FC = (): JSX.Element => {
 			variables: {
 				theme: 'aaa'
 			},
-			refetchQueries: [{ query: GET_USER }]
+			refetchQueries: [{ query: USER_SETTINGS }]
 		})
 	}
 
 	useEffect(() => {
-		refreshToken()
-			.then(data => {
-				setAccessToken(data.accessToken)
-				setUser(data.accessToken)
-			})
+		refreshToken().then((data) => {
+			setAccessToken(data.accessToken)
+			setUser(data.accessToken)
+		})
 	}, [])
 
 	useEffect(() => {
@@ -102,27 +105,7 @@ const App: React.FC = (): JSX.Element => {
 		}
 	}, [user])
 
-	// useEffect(() => {
-	// 	if (user) {
-	// 		loadUser()
-	// 		refreshToken()
-	// 			.then(data => {
-	// 				setAccessToken(data.accessToken)
-	// 				setUser(data.accessToken)
-	// 			})
-	// 		const token = JSON.parse(window.atob(user.split('.')[1]))
-	// 		const timeout = tokenExpiresIn(token)
-	// 		setTimeout(() => {
-	// 			refreshToken()
-	// 				.then(data => {
-	// 					setAccessToken(data.accessToken)
-	// 					setUser(data.accessToken)
-	// 				})
-	// 		}, timeout - 2000)
-	// 	}
-	// }, [user])
-
-	const theme = userTheme?.user[0].theme || 'light'
+	const theme = userTheme?.userSettings[0].theme ?? 'light'
 
 	return (
 		<ThemeProvider theme={theme === 'light' ? themeLight : themeDark}>
@@ -136,15 +119,9 @@ const App: React.FC = (): JSX.Element => {
 						changeTheme
 					}}
 				>
-					<Routes />
+					<Router />
 				</PlayerContext.Provider>
-				{
-					isPlayerVisible &&
-						<CustomPlayer
-							episode={episode}
-							theme={theme}
-						/>
-				}
+				{isPlayerVisible && <CustomPlayer episode={episode} theme={theme} />}
 			</Wrapper>
 			<GlobalStyle />
 		</ThemeProvider>
