@@ -1,11 +1,10 @@
-
 import React, { useContext } from 'react'
-import { NavLink, withRouter } from 'react-router-dom'
+import { NavLink } from 'react-router-dom'
 import { HeaderStyle, NavStyle, ToolStyle, ThemeSwitch } from './styles/Header'
 
 import { PlayerContext } from '../UseContext'
 
-const Header: React.FC = () => {
+const Header = (): JSX.Element => {
 	const { changeTheme } = useContext(PlayerContext)
 
 	const handleTheme = (): void => {
@@ -17,40 +16,24 @@ const Header: React.FC = () => {
 	return (
 		<HeaderStyle>
 			<NavStyle>
-				<NavLink
-					to="/"
-					exact
-					activeClassName="link-active"
-				>
+				<NavLink to="/" activeClassName="link-active">
 					Home
 				</NavLink>
-				<NavLink
-					to="/search"
-					activeClassName="link-active"
-				>
+				<NavLink to="/search" activeClassName="link-active">
 					Search
 				</NavLink>
-				<NavLink
-					to="/favorites"
-					activeClassName="link-active"
-				>
+				<NavLink to="/favorites" activeClassName="link-active">
 					Favorites
 				</NavLink>
-				<NavLink
-					to="/account"
-					activeClassName="link-active"
-				>
+				<NavLink to="/account" activeClassName="link-active">
 					Account
 				</NavLink>
 			</NavStyle>
 			<ToolStyle>
-				<ThemeSwitch
-					onClick={handleTheme}
-				>
-				</ThemeSwitch>
+				<ThemeSwitch onClick={handleTheme}></ThemeSwitch>
 			</ToolStyle>
-		</HeaderStyle >
+		</HeaderStyle>
 	)
 }
 
-export default (withRouter(Header))
+export default Header
