@@ -1,11 +1,6 @@
-
 import React from 'react'
 import { LoadingStyle } from './styles/Loading'
 
-export const Loading: React.FC = () => {
-	return (
-		<LoadingStyle>
-			...loading data
-		</LoadingStyle>
-	)
+export const Loading = (): JSX.Element => {
+	return <LoadingStyle>...loading data</LoadingStyle>
 }
