@@ -1,23 +1,22 @@
-
 import styled from 'styled-components'
 
 export const EpisodeStyle = styled.div`
-    position: fixed;
+	position: fixed;
 	top: 0;
 	bottom: 0;
 	left: 0;
 	right: 0;
-    background-color: ${(props): string => props.theme.playerBackground};
+	background-color: ${(props): string => props.theme.playerBackground};
 	padding: 1rem;
 	z-index: 10;
 `
 
 export const TopStyle = styled.div`
-    display: flex;
+	display: flex;
 `
 
 export const HeaderStyle = styled.div`
-    margin-bottom: 2rem;
+	margin-bottom: 2rem;
 	display: flex;
 	flex-direction: column;
 	padding-top: 2rem;
@@ -30,14 +29,13 @@ export const TitleStyle = styled.div`
 			margin-right: 1rem;
 		}
 	}
-
 `
 export const ContentStyle = styled.div`
 	overflow-y: scroll;
 `
 
 export const ImageStyle = styled.div`
-    width: 100%;
+	width: 100%;
 	display: flex;
 	align-items: center;
 	justify-content: center;
