@@ -1,14 +1,14 @@
 import styled from 'styled-components'
 
 export const LoadingStyle = styled.div`
-    position: fixed;
-    top: 0%;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    background-color: gainsboro;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    opacity: .9;
+	position: fixed;
+	top: 0%;
+	bottom: 0;
+	left: 0;
+	right: 0;
+	background-color: gainsboro;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	opacity: 0.9;
 `
