@@ -2,5 +2,5 @@ import styled from 'styled-components'
 
 export const ModalStyle = styled.div`
 	width: 90%;
-    height: 50%;
+	height: 50%;
 `
