@@ -1,30 +1,29 @@
-
 export interface Podcast {
-	feedUrl: string;
-	collectionId: string;
-	artworkUrl100: string;
+	feedUrl: string
+	collectionId: string
+	artworkUrl100: string
 }
 
 export interface Episode {
-	id?: number;
-	title: string;
-	description: string;
-	url: string;
-	pubDate: string;
-	duration: string;
-	image?: string;
+	id?: number
+	title: string
+	description: string
+	url: string
+	pubDate: string
+	duration: string
+	image?: string
 }
 
 export interface Favorite {
-	id: number;
-	title: string;
-	description: string;
-	url: string;
-	pubDate: string;
-	duration: string;
-	image?: string;
+	id: number
+	title: string
+	description: string
+	url: string
+	pubDate: string
+	duration: string
+	image?: string
 }
 
 export interface KeyValue {
-	[key: string]: string;
+	[key: string]: string
 }
