@@ -1,27 +1,5 @@
 import { gql } from 'apollo-boost'
 
-export const LOGIN = gql`
-	mutation Login($email: String!, $password: String!) {
-		login(email: $email, password: $password) {
-			accessToken
-		}
-	}
-`
-
-export const REGISTER = gql`
-	mutation Register($email: String!, $password: String!) {
-		register(email: $email, password: $password)
-	}
-`
-
-export const GET_USER = gql`
-	query User {
-		user {
-			theme
-		}
-	}
-`
-
 export const SUBSCRIBE = gql`
 	mutation Subscribe($url: String!) {
 		subscribe(url: $url)
@@ -37,7 +15,7 @@ export const UNSUBSCRIBE = gql`
 export const GET_PODCASTS = gql`
 	query Podcasts {
 		podcasts {
-			id,
+			id
 			url
 		}
 	}
@@ -45,19 +23,19 @@ export const GET_PODCASTS = gql`
 
 export const SET_FAVORITE = gql`
 	mutation Favorite(
-		$title: String!,
-		$description: String!,
-		$url: String!,
-		$duration: String!,
-		$pubDate: String!,
+		$title: String!
+		$description: String!
+		$url: String!
+		$duration: String!
+		$pubDate: String!
 		$image: String!
 	) {
 		setFavorite(
-			title: $title,
-			description: $description,
-			url: $url,
-			duration: $duration,
-			pubDate: $pubDate,
+			title: $title
+			description: $description
+			url: $url
+			duration: $duration
+			pubDate: $pubDate
 			image: $image
 		)
 	}
@@ -75,9 +53,9 @@ export const GET_FAVORITES = gql`
 			id
 			description
 			title
-			url,
-			duration,
-			pubDate,
+			url
+			duration
+			pubDate
 			image
 		}
 	}
@@ -86,10 +64,10 @@ export const GET_FAVORITES = gql`
 export const FETCH_PODCASTS = gql`
 	query FetchPodcasts($urls: [String!]!) {
 		fetchPodcasts(urls: $urls) {
-			url,
-			title,
-			pubDate,
-			description,
+			url
+			title
+			pubDate
+			description
 			image
 		}
 	}
@@ -98,30 +76,51 @@ export const FETCH_PODCASTS = gql`
 export const FETCH_PODCASTS_EPISODES = gql`
 	query FetchPodcastEpisodes($url: String!) {
 		fetchPodcastEpisodes(url: $url) {
-			url,
-			title,
-			pubDate,
-			description,
-			image,
+			url
+			title
+			pubDate
+			description
+			image
 			episodes {
-				title,
-				description,
-				pubDate,
-				url,
+				title
+				description
+				pubDate
+				url
 				duration
 			}
 		}
 	}
 `
 
-export const SET_THEME = gql`
-	mutation setTheme($theme: String!) {
-		setTheme(theme: $theme)
+export const SET_TO_HISTORY = gql`
+	mutation SetToHistory(
+		$title: String!
+		$description: String!
+		$url: String!
+		$duration: String!
+		$pubDate: String!
+		$image: String!
+	) {
+		setToHistory(
+			title: $title
+			description: $description
+			url: $url
+			duration: $duration
+			pubDate: $pubDate
+			image: $image
+		)
 	}
 `
 
-export const LOGOUT = gql`
-	mutation Logout($token: String!) {
-		logout(token: $token)
+export const FETCH_HISTORY = gql`
+	query FetchHistory {
+		fetchHistory {
+			title
+			description
+			url
+			duration
+			pubDate
+			image
+		}
 	}
 `
