@@ -1,21 +1,26 @@
+import React, { useContext, useEffect } from 'react'
+import { useNavigate, Link } from 'react-router-dom'
+import { useMutation, useLazyQuery } from '@apollo/react-hooks'
 
-import React, { useContext } from 'react'
-import { RouteComponentProps } from 'react-router-dom'
-import { useMutation } from '@apollo/react-hooks'
-
-import { LOGOUT } from '../query/query'
+import { LOGOUT, USER_PROFILE } from '../query/user_query'
 import { PlayerContext } from '../UseContext'
 import { getAccessToken } from '../accessToken'
-import { AccountContent } from './styles/Account'
+import { AccountContent, AccountMain, AccountTop } from './styles/Account'
 
-export const Account: React.FC<RouteComponentProps> = ({ history }) => {
+export const Account = (): JSX.Element => {
+	const navigate = useNavigate()
 	const { user } = useContext(PlayerContext)
 
 	const [logout] = useMutation(LOGOUT)
+	const [fetchUserProfile, { data: profile }] = useLazyQuery(USER_PROFILE)
+
+	useEffect(() => {
+		fetchUserProfile()
+	}, [])
 
 	const handleLogout = (): void => {
 		if (!user) {
-			history.push('/login')
+			navigate('/login')
 			window.location.reload()
 			return
 		}
@@ -24,24 +29,23 @@ export const Account: React.FC<RouteComponentProps> = ({ history }) => {
 				token: getAccessToken()
 			}
 		}).then(() => {
-			history.push('/login')
+			navigate('/login')
 			window.location.reload()
 		})
 	}
+
 	return (
 		<AccountContent>
-			<div>
-				Hello user
-			</div>
-			<span
-				onClick={handleLogout}
-			>
-				{
-					user
-						? 'logout'
-						: 'login'
-				}
-			</span>
+			<AccountTop>
+				<h5>{profile?.userProfile.email}</h5>
+				<div>
+					<span onClick={handleLogout}>{user ? 'logout' : 'login'}</span>
+				</div>
+			</AccountTop>
+			<AccountMain>
+				<Link to="/friends">Find friends</Link>
+				<Link to="/history">Listeining history</Link>
+			</AccountMain>
 		</AccountContent>
 	)
 }
