@@ -1,27 +1,26 @@
-
 import React, { useEffect, useContext } from 'react'
 import { useLazyQuery } from '@apollo/react-hooks'
 import { Link } from 'react-router-dom'
-import { GET_PODCASTS, FETCH_PODCASTS } from '../query/query'
+import { GET_PODCASTS, FETCH_PODCASTS } from '../query/podcast_query'
 import { ContentStyle, LoginStyle } from './styles/Home'
 import { PlayerContext } from '../UseContext'
 
 interface Podcast {
-	id: string;
-	url: string;
-	image: string;
+	id: string
+	url: string
+	image: string
 }
 
 interface Subs {
-	image: string;
-	url: string;
+	image: string
+	url: string
 }
 
 interface UrlObj {
-	[key: string]: string;
+	[key: string]: string
 }
 
-export const Home: React.FC = () => {
+export const Home = (): JSX.Element => {
 	const { user } = useContext(PlayerContext)
 
 	const [fetchPodcasts, { data: podcasts }] = useLazyQuery(GET_PODCASTS)
@@ -39,49 +38,32 @@ export const Home: React.FC = () => {
 
 	return (
 		<div>
-			<div>
-				{
-					loading	&&
-						<div>...loading</div>
-				}
-			</div>
+			<div>{loading && <div>...loading</div>}</div>
 			<ContentStyle>
-				{
-					subscriptions?.fetchPodcasts
-						.map((podcast: Podcast, index: number): JSX.Element => (
-							<Link
-								to={{
-									pathname: '/podcast',
-									state: {
-										feedUrl: podcast.url
-									}
-								}}
-								key={index}
-							>
-								<img src={podcast.image} />
-							</Link>
-						))
-				}
+				{subscriptions?.fetchPodcasts.map(
+					(podcast: Podcast, index: number): JSX.Element => (
+						<Link
+							to={{
+								pathname: '/podcast',
+								hash: `#${podcast.url}`
+							}}
+							key={index}
+						>
+							<img src={podcast.image} />
+						</Link>
+					)
+				)}
 			</ContentStyle>
-			{
-				!user &&
-					<LoginStyle>
-						<h1>
-							Podcast
-						</h1>
-						<div>
-							<Link to="/login">
-								login
-							</Link>
-							<span>
-								or
-							</span>
-							<Link to="/register">
-								register
-							</Link>
-						</div>
-					</LoginStyle>
-			}
+			{!user && (
+				<LoginStyle>
+					<h1>Podcast</h1>
+					<div>
+						<Link to="/login">login</Link>
+						<span>or</span>
+						<Link to="/register">register</Link>
+					</div>
+				</LoginStyle>
+			)}
 		</div>
 	)
 }
