@@ -1,17 +1,15 @@
-
 import React, { useState, useEffect } from 'react'
-import { RouteComponentProps } from 'react-router'
 import { Link } from 'react-router-dom'
 
 import { Podcast } from '../models/models'
 import { TitleStyle, SearchStyle, ListStyle, SearchContainerStyle } from './styles/Search'
 
 interface State {
-	data: Podcast[];
-	isLoading: boolean;
+	data: Podcast[]
+	isLoading: boolean
 }
 
-export const Search: React.FC<RouteComponentProps> = (): JSX.Element => {
+export const Search = (): JSX.Element => {
 	const [searchQuery, setSearchQuery] = useState('')
 	const [{ data, isLoading }, setData] = useState<State>({
 		data: [],
@@ -23,15 +21,16 @@ export const Search: React.FC<RouteComponentProps> = (): JSX.Element => {
 			data: [],
 			isLoading: true
 		})
-		window.fetch(`https://itunes.apple.com/search?term=${searchQuery}&entity=podcast`)
-			.then(x => x.json())
-			.then(y => {
+		window
+			.fetch(`https://itunes.apple.com/search?term=${searchQuery}&entity=podcast`)
+			.then((x) => x.json())
+			.then((y) => {
 				setData({
 					data: y.results,
 					isLoading: false
 				})
 			})
-			.catch(error => {
+			.catch((error) => {
 				console.log(error)
 				setData({
 					data: [],
@@ -42,36 +41,26 @@ export const Search: React.FC<RouteComponentProps> = (): JSX.Element => {
 
 	return (
 		<SearchContainerStyle>
-			<TitleStyle>
-				{
-					isLoading
-						? 'Searching...'
-						: 'Search'
-				}
-			</TitleStyle>
+			<TitleStyle>{isLoading ? 'Searching...' : 'Search'}</TitleStyle>
 			<SearchStyle>
-				<input
-					type="text"
-					value={searchQuery}
-					onChange={(ev): void => setSearchQuery(ev.target.value)}
-				/>
+				<input type="text" value={searchQuery} onChange={(ev): void => setSearchQuery(ev.target.value)} />
 			</SearchStyle>
 			<ListStyle>
-				{
-					data.map((item: Podcast): JSX.Element => (
-						<Link
-							to={{
-								pathname: '/podcast',
-								state: {
-									feedUrl: item.feedUrl
-								}
-							}}
-							key={item.collectionId}
-						>
-							<img src={item.artworkUrl100} />
-						</Link>
-					))
-				}
+				{data.map(
+					(item: Podcast): JSX.Element => {
+						return (
+							<Link
+								to={{
+									pathname: `/podcast#`,
+									hash: item.feedUrl
+								}}
+								key={item.collectionId}
+							>
+								<img src={item.artworkUrl100} />
+							</Link>
+						)
+					}
+				)}
 			</ListStyle>
 		</SearchContainerStyle>
 	)
