@@ -1,19 +1,28 @@
-
 import styled from 'styled-components'
 
 export const AccountContent = styled.div`
-	padding: 2rem .5rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    div {
-        margin-bottom: 2rem;
-    }
-    span {
-        padding: .5rem;
+	padding: 2rem 0.5rem;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	> div {
+		width: 100%;
+		margin-bottom: 2rem;
+	}
+	span {
+		padding: 0.5rem;
 		margin: 0 1rem;
 		text-decoration: none;
 		color: white;
 		background-color: blueviolet;
-    }
+	}
+`
+
+export const AccountTop = styled.div`
+	display: flex;
+	align-items: center;
+`
+export const AccountMain = styled.div`
+	display: flex;
+	flex-direction: column;
 `
