@@ -1,9 +1,9 @@
-
 const webpack = require('webpack')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
+const Dotenv = require('dotenv-webpack')
 
 const webpackMerge = require('webpack-merge')
-const modeConfig = env => require(`./build-utils/webpack.${env}`)(env)
+const modeConfig = (env) => require(`./build-utils/webpack.${env}`)(env)
 const presetConfig = require('./build-utils/loadPresets')
 const commonPaths = require('./build-utils/common-paths')
 
@@ -84,6 +84,9 @@ module.exports = ({ mode, presets } = { mode: 'production', presets: [] }) => {
 				}
 			},
 			plugins: [
+				new Dotenv({
+					path: './.env'
+				}),
 				new HtmlWebpackPlugin({
 					title: 'Podcast',
 					template: 'public/index.html',
